Version 0.4.0
Add the calibration when removing the external 32.768kHz oscillator;
Add HCI project binary files;
proj_rcu:
    Add a few functions and fixed a few bugs.

+++ firmware/hal/hal_adc.c
@@ -512,7 +512,7 @@ int hal_adc_measure_temp(adc_dev_t* dev, int32_t* pval)
     return ADC_ERR_OK;
 }
 
-int hal_adc_measure_vbat(adc_dev_t* dev, int16_t* pval)
+int hal_adc_measure_vbat(adc_dev_t* dev, int32_t* pval)
 {
     if (!dev)
         return ADC_ERR_INVALID_PARAM;
